Change: Use simdjson::dom API
After some benchmarking I found that the simdjson::dom API is slightly faster than the
simdjson::ondemand API, likely because it doesn't have to iterate over fields over and over again.
This happens because in the ondemand approach it uses an iterator and will search for the field
from there. The DOM approach allows us to store that order at the start. While I'd love to have a
dynamic approach, it doesn't seem to be possible with the current simdjson API.
This commit also introduces the usage of std::optional to properly indicate that something is not
present, which happens fairly often in glTF files.